Mejora en añadir y paso a float la cantidad pastillas
Mejora en añadir medicamento y en contacto

+++ app/src/main/java/com/example/ixtilion/proyecto/Medicamento.java
@@ -5,17 +5,16 @@ package com.example.ixtilion.proyecto;
  */
 public class Medicamento {
     private String nombre; //clave
-    private int num_pastillas; //numero de pastillas que se posee
+    private float num_pastillas; //numero de pastillas que se posee
 
-    public Medicamento (String nombre, int num_pastillas){
+    public Medicamento (String nombre, float num_pastillas){
         this.nombre=nombre;
         this.num_pastillas=num_pastillas;
     }
     public String getNombre(){
         return this.nombre;
     }
-    public int getNum_pastillas(){
+    public float getNum_pastillas(){
         return this.num_pastillas;
     }
-
 }
